Move which_problem code
